Saved work on research
This was a test branch to see if the NR SeriLong link could be used to obtain the trace.id and span.id - research shows that it doesn't work... being async may have something to do with it. I don;t see this branch eve seeing the light of day.